Client/requirements.txt
mysql-connector-python
schedule
playsound
pygame
